Remove undefined division by zero behaviour in `gmp_klee_inv_arg.x86_64` and adjust spec to expect calling abort instead.
Although the division by zero bug is a real bug a compiler is free to do
whatever it wants due to this being undefined behaviour.

It has been observed that running some LLVM passes removes the operation
and at execution time this causes the `abort()` immediately after to be
called.

Although it would be nice to have a "division by zero bug" in our
benchmark suite it has been agreed that in this case reaching the
`abort()` should be allowed and therefore we need to remove the
division by zero operation.